Refactoring index.tsx. Renamed fetchFunctions.tsx to jsonServerFunctions.tsx. Changed import fetchFunctions to jsonServerFunctions. Moved the jsonServerFunctions.tsx to utils folder and deleted json-server folder inside the utils folder
